pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visionCompare"
version = "1.0.0"
description = "基于PySide6的图像对比工具"
readme = "README.md"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "PySide6>=6.0.0",
    "Pillow>=8.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black",
    "flake8"
]

[project.scripts]
visioncompare = "src.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']